requirements.txt
appdirs==1.4.3
certifi==2017.1.23
future==0.16.0
packaging==16.8
pyparsing==2.2.0
python-telegram-bot==5.3.0
six==1.10.0
transmissionrpc==0.11
urllib3==1.20
